Interactive plotting workflow with saved state in basic_testing.py.
